finalize final

+++ src/main/java/et/com/movieReview/dto/ResponseDto/ReviewResponse.java
@@ -8,6 +8,6 @@ import java.util.List;
 @Builder
 public class ReviewResponse {
     private List<ReviewResponseDto> reviewList;
-    private Integer totalResult;
+    private Long totalResult;
     private String status;
 }
